+++ src/Symfony/Component/Security/Http/CHANGELOG.md
@@ -14,6 +14,7 @@ CHANGELOG
  * Add `ImpersonateUrlGenerator::generateImpersonationForm()` and `generateExitForm()` to build a form that switches the user with a POST
  * Add `$targetUri` argument to `ImpersonateUrlGenerator::generateImpersonationPath()` and `generateImpersonationUrl()`
  * Configure the decorated handler of `CustomAuthenticationSuccessHandler` and `CustomAuthenticationFailureHandler` when they are called instead of when they are built, so that a single handler can be shared by several authenticators
+ * Add `BeforeAuthenticateEvent`, dispatched before invoking an authenticator (useful for auditing/logging authentication attempts)
 
 8.1
 ---